pyproject.toml
[project]
name = "services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiopg==1.4.0",
    "aiosignal==1.3.2",
    "aiosqlite==0.20.0",
    "alembic==1.14.0",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "asttokens==3.0.0",
    "async-timeout==4.0.3",
    "asyncpg==0.30.0",
    "attrs==24.3.0",
    "awslambdaric==3.0.2",
    "backcall==0.2.0",
    "bcrypt==4.2.1",
    "beautifulsoup4==4.12.3",
    "bleach==6.2.0",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "cryptography==44.0.0",
    "databases==0.9.0",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "deprecation==2.1.0",
    "dnspython==2.7.0",
    "docopt==0.6.2",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "executing==2.1.0",
    "fastapi==0.115.6",
    "fastapi-cli==0.0.7",
    "fastjsonschema==2.21.1",
    "filetype==1.2.0",
    "frozenlist==1.5.0",
    "gotrue==2.11.1",
    "graphql-core==3.2.6",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "h2==4.1.0",
    "hpack==4.0.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "hyperframe==6.0.1",
    "idna==3.10",
    "ipython==8.12.3",
    "jedi==0.19.2",
    "jinja2==3.1.5",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyterlab-pygments==0.3.0",
    "mako==1.3.8",
    "mangum==0.19.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.1.0",
    "multidict==6.1.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.5",
    "nbformat==5.10.4",
    "packaging==24.2",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "passlib==1.7.4",
    "pexpect==4.9.0",
    "phonenumbers==9.0.3",
    "pickleshare==0.7.5",
    "pipreqs==0.5.0",
    "platformdirs==4.3.6",
    "postgrest==0.19.1",
    "prompt-toolkit==3.0.48",
    "propcache==0.2.1",
    "psycopg2==2.9.10",
    "psycopg2-binary==2.9.10",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "pydantic-extra-types==2.10.3",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.20",
    "pywin32==310",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "realtime==2.1.0",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rich==13.9.4",
    "rich-toolkit==0.12.0",
    "rpds-py==0.22.3",
    "rsa==4.9",
    "shellingham==1.5.4",
    "simplejson==3.20.1",
    "six==1.17.0",
    "snapshot-restore-py==1.0.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.36",
    "sqlmodel==0.0.22",
    "stack-data==0.6.3",
    "starlette==0.41.3",
    "storage3==0.11.1",
    "strawberry-graphql==0.266.0",
    "strenum==0.4.15",
    "supabase==2.12.0",
    "supafunc==0.9.2",
    "tinycss2==1.4.0",
    "tornado==6.4.2",
    "traitlets==5.14.3",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "watchfiles==1.0.3",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "websockets==13.1",
    "yarg==0.1.9",
    "yarl==1.18.3",
]
